Split calendar.js into routes/calendar/ sub-routers (Chunk 9)
Pure move, no behaviour change. calendar.js (945 lines) split into
routes/calendar/ (events.js, openEvents.js, eventMembers.js, shared
helpers.js, index.js). events.js holds the read side and defines its
literal /events/pdf, /events/excel, /events/search routes before
/events/:eventId so the param does not capture them. openEvents.js owns
the non-group /open-events CRUD; eventMembers.js owns the attendance
sub-resource; helpers.js holds the fmtDateUK/fmtTime formatters shared by
events.js and eventMembers.js. app.js import updated.

Completes Chunk 9 (all seven oversized backend route files split).
Updates ImprovementPlan.md, CHANGELOG.md, and eventFilters.js comments.
